refactor: standard fastapi project structure for better main… (#217)
* refactor: maintain standard fastapi project structure for better maintainability


* chore: merge main into branch


* fix: run pre-commit


* fix: setup.py


* fix: service discovery time import


* fix: service discovery time import


* fix: delete unused file


* fix: merge main and fix log_stats.py


* chore: run pre-commit


* fix: correct experimental imports


* fix: parser error


* fix: experimental feature flag


* fix: wrapper call issue


* fix: add missing lifespan


* chore: add TODO comment


---------
