--- src/vllm_router/files/__init__.py
@@ -1,10 +1,0 @@
-from vllm_router.files.file_storage import FileStorage
-from vllm_router.files.files import OpenAIFile
-from vllm_router.files.storage import Storage, initialize_storage
-
-__all__ = [
-    "OpenAIFile",
-    "Storage",
-    "FileStorage",
-    "initialize_storage",
-]
