+++ src/vllm_router/perf-test.sh
@@ -8,7 +8,7 @@ fi
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 # Run router.py from the correct directory
-python3 "$SCRIPT_DIR/router.py" --port "$1" \
+python3 "$SCRIPT_DIR/app.py" --port "$1" \
     --service-discovery static \
     --static-backends "http://localhost:9004,http://localhost:9001,http://localhost:9002,http://localhost:9003" \
     --static-models "fake_model_name,fake_model_name,fake_model_name,fake_model_name" \
