+++ src/vllm_router/protocols.py
@@ -1,3 +1,4 @@
+import time
 from typing import List, Optional
 
 from pydantic import BaseModel, ConfigDict, Field, model_validator
