+++ src/vllm_router/services/files_service/__init__.py
@@ -1,0 +1,8 @@
+from vllm_router.services.files_service.storage import Storage, initialize_storage
+
+__all__ = [
+    "OpenAIFile",
+    "Storage",
+    "FileStorage",
+    "initialize_storage",
+]
